Initial commit - highlighter class, tests, and build config
The implementation is in the `Highlighter` class. Its function is to
match text in an HTML source, and wrap any matching text in a <span/>
element which is decorated with a well-defined class name; this allows
styling to highlight matches.  `Highlighter` skips over tags, and can
match text on a page that is separated in the DOM by tags.

Implementation is in typescript, and tests are executed via the browser
using mocha.